+++ ChangeLog.md
@@ -19,6 +19,7 @@
     - Let a role be used several times inside of the roles hierarchy
     - Define new Role.isRoleScoped() function
     - Improve Roles.current() content, adding scoped and globals categories
+    - Define new Roles.removeUserAssignmentsForRoles() function
 
 ### 1.2.0